+++ CHANGES.md
@@ -6,7 +6,7 @@
 - new commands: --get-service-containers:service, --get-number-of-containers:service, --get-deploy-replicas:service
 - AbstractManager class
 - refactoring calling docker and docker compose commands
-- --help command print all available commands
+- --help prints all available commands
 
 ## v15.0.0 (2024-02-23)
 - README.md